README.md: Fix lookup example
Adjust one of the IP addresses in the slice used to build the set, in
order to fix one lookup example. Namely, when "1.1.2.0/24", it should be
the result for the lookup for "1.1.2.2", having a longer matching prefix
than "1.1.0.0/20".

We could fix the expected value in the assertion, but instead we align
the contents of the set with the example in examples/lookup.rs.

Fixes: c253ebdc9f5e ("deep refactoring")
Signed-off-by: Quentin Monnet <qmo@qmon.net>